% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting5jars-package.R
\docType{package}
\name{melting5jars}
\alias{melting5jars}
\alias{melting5jars-package}
\title{Java Archive Wrapper Supporting the 'melting5' Package}
\description{
Java Archive Wrapper Supporting the 'melting5' Package
}
\author{
Bob Rudis (bob@rud.is)
}
